ai_social_bot/requirements.txt
requests
python-dotenv
flask
openai
langdetect
pandas
scikit-learn
schedule
textblob
chromadb
